+++ Pandora/Paradigm/Algebraic/Functor.hs
@@ -18,7 +18,7 @@ import Pandora.Pattern.Functor.Adjoint (Adjoint ((-|), (|-)))
 import Pandora.Pattern.Operation.One (One (One))
 import Pandora.Pattern.Operation.Zero (Zero, absurd)
 import Pandora.Pattern.Operation.Exponential (type (-->), type (--<), (&))
-import Pandora.Paradigm.Algebraic.Product ((:*:) ((:*:)))
+import Pandora.Pattern.Operation.Product ((:*:) ((:*:)))
 import Pandora.Paradigm.Algebraic.Sum ((:+:) (Option, Adoption))
 import Pandora.Paradigm.Primary.Functor.Proxy (Proxy (Proxy))