+++ Pandora/Paradigm/Primary/Transformer/Continuation.hs
@@ -11,7 +11,7 @@ import Pandora.Pattern.Functor.Monad (Monad)
 import Pandora.Pattern.Transformation.Liftable (Liftable (lift))
 import Pandora.Core.Interpreted (Interpreted (Primary, run, unite, (<~)))
 import Pandora.Pattern.Operation.Exponential ((%), type (-->))
-import Pandora.Paradigm.Algebraic.Product ((:*:))
+import Pandora.Pattern.Operation.Product ((:*:))
 import Pandora.Paradigm.Algebraic (point)
 
 newtype Continuation r t a = Continuation ((->) ((->) a (t r)) (t r))